Break out 686c utils into separate files
